Change Double -> Float

+++ src/main/scala/com/raphtory/dev/blockchain/graphbuilders/chab_C2C_GB.scala
@@ -22,9 +22,9 @@ class chab_C2C_GB extends GraphBuilder[String]{
         srcClusterId,
         dstClusterId,
         Properties(
-          DoubleProperty("BitCoin", btcAmount),
-          DoubleProperty("USD", usdAmount),
-          DoubleProperty("Transaction", transactionId)
+          FloatProperty("BitCoin", btcAmount),
+          FloatProperty("USD", usdAmount),
+          LongProperty("Transaction", transactionId)
         ),
         Type("Transfer")
       )
@@ -34,12 +34,12 @@ class chab_C2C_GB extends GraphBuilder[String]{
   //converts the line into a case class which has all of the data via the correct name and type
   def formatLine(line: Array[String]): Datapoint =
     Datapoint(
-      line(1).toDouble / 100000000, 			//Amount of transaction in BTC
+      line(1).toFloat / 100000000, 			//Amount of transaction in BTC
       line(2).toLong, 			            //ID of destination cluster
       line(3).toLong, 			            //ID of source cluster
       line(4).toLong * 1000,            //Time of transaction in seconds (milli in Raph)
       line(5).toLong, 			            //ID of transaction, can be similar for many records
-      line(6).toDouble / 100000 			    //Amount of transaction in USD
+      line(6).toFloat / 100000 			    //Amount of transaction in USD
 
     )
 
@@ -47,12 +47,11 @@ class chab_C2C_GB extends GraphBuilder[String]{
 
 
   case class Datapoint(
-                        amount: Double,       //Amount of transaction in Satoshi
+                        amount: Float,       //Amount of transaction in Satoshi
                         dstCluster: Long,   //ID of destination cluster
                         srcCluster: Long,   //ID of source cluster
                         time: Long,         //Time of transaction in seconds
                         txid: Long,          //ID of transaction, can be similar for many records
-                        usd: Double          //Amount of transaction in USD
+                        usd: Float          //Amount of transaction in USD
                       )
-
 }
